move BufferedReader out of nu-command
